fix: reduce duplication after PR comments
